market registry mapping
